Fixes for out-of order GTFs and protein_coding
Instead of pushing transcripts onto a region array when we start
processing the next one, leave all of that to the end of a
chromosome block, this deals with the issue of undefined
transcripts and is much neater generally.

Also, if a user's input gives a transcript biotype as protein
coding but does not provide and CDS lines, default it to being a
psuedogene and remove the translation.