Change handling of refseq cache
Users now have to specify --refseq along with --cache or --offline
to use the refseq cache. '_refseq' will be appended to the cache
dir, and refseq caches will be built with that suffix in their path.

The installer has also been updated to cope with this. There may
well be helpdesk/dev questions about this, but this solution is much
neater than that which exists now!!!